==> docs/html/.doctrees/nbsphinx/files/notebooks/4.0-cbw-app-train.ipynb <==
# Model training application entry point

import subprocess
import pandas as pd

## Pre-processing both train and test CSVs in data/external and save them in data/processed

%run ../data/2.0-cbw-data-preprocessing.ipynb

processed_train = pd.read_csv('../../data/processed/processed_application_train.csv')
processed_test = pd.read_csv('../../data/processed/processed_application_test.csv')

## Display pre-processed train and test CSVs first lines (head)

processed_train.head()

processed_test.head()

## Train models and track them with MLFlow

def train_models():
    subprocess.run('mlflow run ../../src/models/train/gradient_boosting')
    subprocess.run('mlflow run ../../src/models/train/xgboost ')
    subprocess.run('mlflow run ../../src/models/train/random_forest')

train_models()

## Running MLFow UI server to visualize runs

subprocess.run('mlflow ui')
