==> docs/html/.doctrees/nbsphinx/files/notebooks/5.0-cbw-app-predict.ipynb <==
# Model prediction application entry point

from src.models import predict_model as pm

## Choosing model for prediction (between "gb", "xgb" and "rf")

model = 'rf'

## Specifying test CSV file path for predictions

file_path = '../../data/processed/processed_application_test.csv'

## Display first 10 predicted results

predictions = pm.get_predictions(model, file_path)
print(predictions[:10])
